# server/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API that retrieves information about clients and their vehicles.",
        "title": "Starter Project",
        "termsOfService": "TBD",
        "contact": {
            "name": "Byron Ojua-Nice",
            "url": "http://firstlaunch.dev",
            "email": "byronojua@firstlaunch.dev"
        },
        "license": {
            "name": "TBD",
            "url": "TBD"
        },
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/clients": {
            "get": {
                "description": "Get all clients and the number of vehicles they have",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Get all clients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.getAllClientsOutput"
                            }
                        }
                    },
                    "404": {
                        "description": "Clients not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "description": "Get a client by their ID and the number of vehicles they have.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Get client by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID (name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.getClientByIdOutput"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    }
                }
            }
        },
        "/clients/{id}/vehicles": {
            "get": {
                "description": "Gets all vehicles for a client by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Get client vehicles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID (name)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.getClientVehiclesOutput"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "404": {
                        "description": "Vehicles not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    }
                }
            }
        },
        "/vehicles/{id}": {
            "get": {
                "description": "Get a vehicle by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get vehicle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vehicle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.getVehicleByIdOutput"
                        }
                    },
                    "400": {
                        "description": "Bad request - Id is required",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    },
                    "404": {
                        "description": "Vehicle not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BasicVehicle": {
            "type": "object",
            "properties": {
                "largest_weight": {
                    "description": "The largest weight the vehicle has carried",
                    "type": "integer"
                },
                "mileage": {
                    "description": "The mileage of the vehicle",
                    "type": "integer"
                },
                "vin": {
                    "description": "The VIN of the vehicle",
                    "type": "string"
                }
            }
        },
        "api.ClientInfo": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "description": "The email of the contact person for the client",
                    "type": "string"
                },
                "contact_name": {
                    "description": "The name of the contact person for the client",
                    "type": "string"
                },
                "name": {
                    "description": "The name of the client",
                    "type": "string"
                },
                "number_of_vehicles": {
                    "description": "The number of vehicles the client has",
                    "type": "integer"
                }
            }
        },
        "api.ErrorsResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Array of error messages",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.VehicleInfo": {
            "type": "object",
            "properties": {
                "client_name": {
                    "type": "string"
                },
                "contact_email": {
                    "type": "string"
                },
                "contact_name": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "vin": {
                    "type": "string"
                },
                "weights": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "api.getAllClientsOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The list of clients",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ClientInfo"
                    }
                }
            }
        },
        "api.getClientByIdOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The client",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.ClientInfo"
                        }
                    ]
                }
            }
        },
        "api.getClientVehiclesOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The vehicle",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.BasicVehicle"
                    }
                }
            }
        },
        "api.getVehicleByIdOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.VehicleInfo"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use the format: `Bearer \u003cyour_token\u003e` (Bearer must be added before the token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}